requirements.txt
asgiref==3.7.2
astroid==3.0.3
atomicwrites==1.4.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.12
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.8.0rc2
dill==0.3.8
Django==3.2
django-filter==23.5
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
flake8==7.0.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.2
platformdirs==4.2.0
pluggy==0.13.1
py==1.11.0
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.8.0
pylint==3.0.3
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.26.0
requests-oauthlib==2.0.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.18
